ashmem: Fix lockdep warning for dynamically allocated mutex lock
While using the mmap_lock mutex lock in ashmem_mmap(), lockdep gives the
following warning:

[   10.531472] INFO: trying to register non-static key.
[   10.531476] The code is fine but needs lockdep annotation, or maybe
[   10.531477] you didn't initialize this object before use?
[   10.531478] turning off the locking correctness validator.
[   10.531482] CPU: 5 PID: 1091 Comm: diag-router Tainted: G        W         5.4.289-Scarlet-v2.2-beta2 #1
[   10.531483] Hardware name: redwood based Qualcomm Technologies, Inc. SM7325 (DT)
[   10.531484] Call trace:
[   10.531492] dump_backtrace+0x0/0x1c0
[   10.531497] dump_stack+0x90/0xcc
[   10.531500] register_lock_class+0x3a8/0x940
[   10.531502] __lock_acquire.isra.0+0x2f0/0x860
[   10.531503] lock_acquire+0xd0/0x200
[   10.531505] __mutex_lock+0xa4/0x944
[   10.531507] mutex_lock_nested+0x20/0x28
[   10.531510] ashmem_mmap+0xc8/0x1e0
[   10.531513] mmap_region+0x368/0x5e0
[   10.531515] do_mmap+0x28c/0x560
[   10.531516] vm_mmap_pgoff+0xe8/0x140
[   10.531518] ksys_mmap_pgoff+0x84/0xc0
[   10.531520] __arm64_sys_mmap+0x2c/0x40
[   10.531521] do_el0_svc+0x68/0x120
[   10.531522] el0_sync_handler+0x11c/0x1c0
[   10.531524] el0_sync+0x140/0x180

This is because the mutex lock was not properly initialized for dynamic
allocation. In ashmem_open(), mmap_lock was initialized using
__MUTEX_INITIALIZER(), which is only safe for static mutex lock declarations and
inadequate for dynamic allocations.

Therefore, register a runtime mutex_init() instead of the static initializer. To
do so, we get rid of the compound literal brace-initializer for "struct tmp" and
explicitly define mmap_lock and prot_mask.

But in the struct, every member that aren't named and all padding bits are
zeroed. Getting rid of it requires the slab cache allocation to be zeroed
initially. That's why, switch to kmem_cache_zalloc() for that which uses
SLAB_ZERO.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>